+++ src/ai-agent-framework/contracts/multi-agent-coordination-profile.js
@@ -1,0 +1,63 @@
+export class MultiAgentCoordinationProfile {
+  constructor({
+    coordinationName,
+    selectionCriteria = [], inappropriateCriteria = [], coordinationGoals = [], roles = [],
+    patterns = [], topologies = [], sharedGoalFields = [], taskOwnershipFields = [],
+    delegationContractFields = [], authorityAttenuationDimensions = [], delegationAdmissionChecks = [],
+    delegationChainFields = [], delegationLifecycleStates = [], delegateSelectionFactors = [],
+    messageTypes = [], messageEnvelopeFields = [], messageAuthorizationChecks = [], sharedStateFields = [],
+    concurrencyRequirements = [], barrierFields = [], conflictTypes = [], conflictResolutionOrder = [],
+    independentReviewCriteria = [], delegateResultFields = [], aggregationRequirements = [],
+    failurePropagationPolicies = [], cancellationFlowSteps = [], humanSupervisionActions = [],
+    humanTakeoverSteps = [], securityThreats = [], evidenceFields = [], observabilitySignals = [],
+    evaluationMeasures = [], operations = [], qualityAttributes = [], architecturalRules = [],
+    architectureBoundaries = [], optionalComposition = true, smallestAgentSet = true,
+    participantsIdentified = true, sharedAuthorizedGoal = true, oneOwnerPerTask = true,
+    explicitDelegationContracts = true, authorityAttenuated = true, noCredentialTransfer = true,
+    initiatorPrincipalPreserved = true, tenantPropertyScopePreserved = true, boundedDepthFanout = true,
+    circularDelegationPrevented = true, versionedMessageContracts = true, authorizedMessagesContext = true,
+    messagesTreatedByTrust = true, sharedStateOutsideConversation = true, systemOfRecordOwnershipPreserved = true,
+    delegateResultsValidated = true, attributionPreserved = true, majorityNotTruth = true,
+    failurePropagationDefined = true, cancellationSuspensionPropagated = true, childLimitsCountAgainstParent = true,
+    humanSupervisionSupported = true, coordinationEvidenceProduced = true, workflowForDurableCoordination = true,
+    modelFrameworkNeutral = true, multiAgentDefault = false, complexityAsIntelligence = false,
+    rolesGrantAuthority = false, falseIndependence = false, unrestrictedPeerMesh = false,
+    ambiguousTaskOwnership = false, delegationImpersonates = false, delegationExpandsAuthority = false,
+    delegationTransfersCredentials = false, unboundedSubcontracting = false, unregisteredDelegates = false,
+    messagesAsAuthority = false, fullParentContextShared = false, transcriptOnlySharedState = false,
+    agentsOverwriteSystemOfRecord = false, majorityAsTruth = false, fabricatedConsensus = false,
+    childStateErasedByCoordinatorFailure = false, delegationMultipliesResources = false,
+    crossTenantDelegationByDefault = false, competingActionsDuringTakeover = false,
+    selectsCommunicationProtocol = false, selectsFrameworkProduct = false
+  }) {
+    this.coordinationName = coordinationName;
+    for (const [key, value] of Object.entries({
+      selectionCriteria, inappropriateCriteria, coordinationGoals, roles, patterns, topologies,
+      sharedGoalFields, taskOwnershipFields, delegationContractFields, authorityAttenuationDimensions,
+      delegationAdmissionChecks, delegationChainFields, delegationLifecycleStates, delegateSelectionFactors,
+      messageTypes, messageEnvelopeFields, messageAuthorizationChecks, sharedStateFields,
+      concurrencyRequirements, barrierFields, conflictTypes, conflictResolutionOrder,
+      independentReviewCriteria, delegateResultFields, aggregationRequirements, failurePropagationPolicies,
+      cancellationFlowSteps, humanSupervisionActions, humanTakeoverSteps, securityThreats, evidenceFields,
+      observabilitySignals, evaluationMeasures, operations, qualityAttributes, architecturalRules,
+      architectureBoundaries
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      optionalComposition, smallestAgentSet, participantsIdentified, sharedAuthorizedGoal, oneOwnerPerTask,
+      explicitDelegationContracts, authorityAttenuated, noCredentialTransfer, initiatorPrincipalPreserved,
+      tenantPropertyScopePreserved, boundedDepthFanout, circularDelegationPrevented, versionedMessageContracts,
+      authorizedMessagesContext, messagesTreatedByTrust, sharedStateOutsideConversation,
+      systemOfRecordOwnershipPreserved, delegateResultsValidated, attributionPreserved, majorityNotTruth,
+      failurePropagationDefined, cancellationSuspensionPropagated, childLimitsCountAgainstParent,
+      humanSupervisionSupported, coordinationEvidenceProduced, workflowForDurableCoordination,
+      modelFrameworkNeutral, multiAgentDefault, complexityAsIntelligence, rolesGrantAuthority,
+      falseIndependence, unrestrictedPeerMesh, ambiguousTaskOwnership, delegationImpersonates,
+      delegationExpandsAuthority, delegationTransfersCredentials, unboundedSubcontracting, unregisteredDelegates,
+      messagesAsAuthority, fullParentContextShared, transcriptOnlySharedState, agentsOverwriteSystemOfRecord,
+      majorityAsTruth, fabricatedConsensus, childStateErasedByCoordinatorFailure, delegationMultipliesResources,
+      crossTenantDelegationByDefault, competingActionsDuringTakeover, selectsCommunicationProtocol,
+      selectsFrameworkProduct
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}